requirements.txt
absl-py==1.4.0
alembic==1.7.7
apgl==0.8.1
appdirs==1.4.4
ase==3.22.1
astor==0.8.1
async-generator==1.10
attrs==22.2.0
autopage==0.5.1
axial-positional-embedding==0.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
brotlipy==0.7.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.0
charset-normalizer==2.0.12
click==8.0.4
cliff==3.10.1
cloudpickle==2.2.1
cmaes==0.9.0
cmd2==2.4.3
colorama==0.4.5
colorlog==6.7.0
config==0.5.1
contextlib2==21.6.0
cryptography @ file:///tmp/build/80754af9/cryptography_1635366128178/work
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dgl-cu102==0.6.1
docker-pycreds==0.4.0
dppy==0.3.2
easydict==1.10
einops==0.4.1
ema-pytorch==0.2.2
entrypoints==0.4
faiss-gpu==1.7.2
fastdtw==0.3.4
filelock==3.3.2
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.5.3
gdown==4.7.1
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.18
google-pasta==0.2.0
googledrivedownloader==0.4
greenlet==2.0.2
grpcio==1.48.2
h5py==3.1.0
hyperopt==0.1.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
iopath==0.1.10
ipdb==0.13.13
ipython==7.16.3
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
json-tricks==3.16.1
json5==0.9.11
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
littleutils==0.2.2
llvmlite==0.36.0
local-attention==1.4.4
Mako==1.1.6
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.3.0
munkres==1.1.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.6.0
networkx==2.5.1
nni==2.6.1
node2vec==0.4.3
numba==0.53.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
ogb==1.3.6
opt-einsum==3.3.0
optuna==3.0.6
outdated==0.2.2
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.1
parso==0.7.1
pathtools==0.1.2
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plyfile==0.8
portalocker==2.7.0
prettytable==2.5.0
product-key-memory==0.1.10
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.2
Pygments==2.14.0
pymongo==4.1.1
pynvml==11.5.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.7
pyperclip==1.8.2
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305763431/work
python-dateutil==2.8.2
python-louvain==0.16
PythonWebHDFS==0.2.3
pytz==2022.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==25.1.2
rdflib==5.0.0
reformer-pytorch==1.4.4
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
responses==0.17.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
schema==0.7.5
SciencePlots==2.1.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
sentry-sdk==1.18.0
setproctitle==1.2.3
simplejson==3.18.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smart-open==6.3.0
smmap==5.0.0
snap==0.5
soupsieve==2.3.2.post1
SQLAlchemy==1.4.49
stevedore==3.5.2
sympy==1.9
tabulate==0.8.10
tensorboard==2.0.0
tensorboardX==2.6
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
testpath==0.6.0
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2020.9.3
tomli==1.2.3
torch==1.9.1+cu102
torch-cluster==1.5.9
torch-ema==0.3
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchaudio==0.9.1
torchdiffeq==0.2.3
torchmetrics==0.8.2
torchvision==0.10.1+cu102
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
typeguard==2.13.3
typing_extensions==4.1.1
urllib3==1.26.11
wandb==0.14.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.3
wrapt==1.15.0
yacs==0.1.8
zipp==3.6.0
